Rename Marketplace Plugin to Marketplace Suite and update configs/tests
