--- src/Set/SyliusMarketplace.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Sylius\SyliusRector\Set;
-
-final class SyliusMarketplace
-{
-    public const SYLIUS_MARKETPLACE = __DIR__ . '/../../config/sets/sylius/marketplace/marketplace-plugin.php';
-}
