+++ src/Set/SyliusPlus.php
@@ -7,10 +7,18 @@ namespace Sylius\SyliusRector\Set;
 final class SyliusPlus
 {
     public const B2B_SUITE = __DIR__ . '/../../config/sets/sylius/plus/b2b-suite.php';
+
+    public const MARKETPLACE_SUITE = __DIR__ . '/../../config/sets/sylius/plus/marketplace-suite.php';
+
     public const LOYALTY_PLUGIN = __DIR__ . '/../../config/sets/sylius/plus/loyalty-plugin.php';
+
     public const MULTI_SOURCE_INVENTORY_PLUGIN = __DIR__ . '/../../config/sets/sylius/plus/multi-source-inventory-plugin.php';
+
     public const MULTI_STORE_PLUGIN = __DIR__ . '/../../config/sets/sylius/plus/multi-store-plugin.php';
+
     public const PLUS_RBAC_PLUGIN = __DIR__ . '/../../config/sets/sylius/plus/plus-rbac-plugin.php';
+
     public const RETURN_PLUGIN = __DIR__ . '/../../config/sets/sylius/plus/return-plugin.php';
+
     public const UPGRADE_TO_MODULAR = __DIR__ . '/../../config/sets/sylius/plus/up-to-plus-modular.php';
 }
