+++ tests/Set/SyliusPlus/MarketplaceSuite/Fixture/class_extending_channel_repository.php.inc
