+++ tests/Set/SyliusPlus/MarketplaceSuite/Fixture/class_extending_product_model.php.inc
