+++ tests/Set/SyliusPlus/MarketplaceSuite/Fixture/class_extending_shop_user_model.php.inc
