+++ tests/Set/SyliusPlus/MarketplaceSuite/config/configured_rule.php
@@ -3,11 +3,10 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Sylius\SyliusRector\Set\SyliusMarketplace;
 use Sylius\SyliusRector\Set\SyliusPlus;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->import(__DIR__ . '/../../../../config/config.php');
-    $rectorConfig->sets([SyliusMarketplace::SYLIUS_MARKETPLACE]);
+    $rectorConfig->import(__DIR__ . '/../../../../../config/config.php');
+    $rectorConfig->sets([SyliusPlus::MARKETPLACE_SUITE]);
     $rectorConfig->importNames();
 };
